lib/datasets/datasets_service/requirements.txt
jsonschema==3.2.0
connexion==2.7.0
SQLAlchemy==1.3.0
tornado==6.0.1
decorator==4.3.2
openapi-core==0.11.0
pyyaml>=4.2b1
swagger-ui-bundle==0.0.5
pronto==1.1.3

pytest-cov==2.7.1
coverage==4.5.3
pytest==4.5.0
pytest-runner==4.4

uwsgi==2.0.18


# jsonschema[format] requires the following otherwise we get an error:
rfc3987==1.3.8
strict-rfc3339==0.7
webcolors==1.8.1
Sphinx==2.3.1
sphinx-rtd-theme==0.4.3
Werkzeug==1.0.1
